Conflicts:
	Math/CNTKMathTest/MatrixSparseDenseInteractionsTests.cpp

+++ Math/Math/CPUMatrix.cpp
@@ -8,7 +8,7 @@
 
 #include "stdafx.h"
 #include "Basics.h"
-#include "fileutil.h"
+#include "File.h"
 
 #include <assert.h>
 #include <stdexcept>
@@ -497,7 +497,7 @@ namespace Microsoft { namespace MSR { namespace CNTK {
     CPUMatrix<ElemType> CPUMatrix<ElemType>::Diagonal() const
     {
         if (m_numRows != m_numCols)
-            LogicError("Diagonal can be called only for square matrix. (rows=%d, cols=%d)", m_numRows, m_numCols);
+            LogicError("Diagonal can be called only for square matrix. (rows=%d, cols=%d)", (int)m_numRows, (int)m_numCols);
 
         CPUMatrix<ElemType> diag(1, m_numCols);
 
@@ -5509,6 +5509,7 @@ namespace Microsoft { namespace MSR { namespace CNTK {
         return *this;
     }
 
+    // note: this function does not depend on the <ElemType> parameter
     template<class ElemType>
     int CPUMatrix<ElemType>::SetNumThreads(int numThreads)
     {
